+++ src/main/java/net/geforcemods/securitycraft/containers/ContainerProjector.java
@@ -1,39 +1,43 @@
 package net.geforcemods.securitycraft.containers;
 
 import net.geforcemods.securitycraft.SCContent;
+import net.geforcemods.securitycraft.SecurityCraft;
+import net.geforcemods.securitycraft.inventory.TileEntityInventoryWrapper;
+import net.geforcemods.securitycraft.network.server.SyncProjector;
 import net.geforcemods.securitycraft.tileentity.TileEntityProjector;
 import net.geforcemods.securitycraft.util.BlockUtils;
+import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.entity.player.InventoryPlayer;
-import net.minecraft.inventory.Container;
 import net.minecraft.inventory.Slot;
 import net.minecraft.item.ItemBlock;
 import net.minecraft.item.ItemStack;
 
-public class ContainerProjector extends Container {
+public class ContainerProjector extends ContainerStateSelectorAccess {
 	public static final int SIZE = 1;
 	private final TileEntityProjector te;
+	private Slot projectedBlockSlot;
 
 	public ContainerProjector(InventoryPlayer inventory, TileEntityProjector te) {
 		this.te = te;
 
-		for (int y = 0; y < 3; y++) {
-			for (int x = 0; x < 9; ++x) {
-				addSlotToContainer(new Slot(inventory, x + y * 9 + 9, 8 + x * 18, 84 + y * 18 + 59));
-			}
-		}
-
-		for (int x = 0; x < 9; x++) {
-			addSlotToContainer(new Slot(inventory, x, 8 + x * 18, 142 + 59));
-		}
-
 		// A custom slot that prevents non-Block items from being inserted into the projector
-		addSlotToContainer(new Slot(te, 36, 79, 23) {
+		projectedBlockSlot = addSlotToContainer(new Slot(new TileEntityInventoryWrapper<>(te, this), 36, 79, 23) {
 			@Override
 			public boolean isItemValid(ItemStack stack) {
 				return stack.getItem() instanceof ItemBlock;
 			}
 		});
+
+		for (int y = 0; y < 3; y++) {
+			for (int x = 0; x < 9; ++x) {
+				addSlotToContainer(new Slot(inventory, x + y * 9 + 9, 8 + x * 18, 84 + y * 18 + 69));
+			}
+		}
+
+		for (int x = 0; x < 9; x++) {
+			addSlotToContainer(new Slot(inventory, x, 8 + x * 18, 142 + 69));
+		}
 	}
 
 	@Override
@@ -45,12 +49,12 @@ public class ContainerProjector extends Container {
 			ItemStack slotStack = slot.getStack();
 			slotStackCopy = slotStack.copy();
 
-			if (index == 36) {
-				if (!mergeItemStack(slotStack, 0, 36, false))
+			if (index < 1) {
+				if (!mergeItemStack(slotStack, 1, 36, true))
 					return ItemStack.EMPTY;
 			}
-			else {
-				if (!mergeItemStack(slotStack, 36, 37, false))
+			else if (index >= 1) {
+				if (!mergeItemStack(slotStack, 0, 1, false))
 					return ItemStack.EMPTY;
 			}
 
@@ -59,7 +63,7 @@ public class ContainerProjector extends Container {
 			else
 				slot.onSlotChanged();
 
-			if (slotStack.getCount() == slotStack.getCount())
+			if (slotStack.getCount() == slotStackCopy.getCount())
 				return ItemStack.EMPTY;
 
 			slot.onTake(player, slotStack);
@@ -69,7 +73,26 @@ public class ContainerProjector extends Container {
 	}
 
 	@Override
+	public void onStateChange(IBlockState state) {
+		te.setProjectedState(state);
+		detectAndSendChanges();
+
+		if (te.getWorld().isRemote)
+			SecurityCraft.network.sendToServer(new SyncProjector(te.getPos(), state));
+	}
+
+	@Override
 	public boolean canInteractWith(EntityPlayer player) {
 		return BlockUtils.isWithinUsableDistance(te.getWorld(), te.getPos(), player, SCContent.projector);
+	}
+
+	@Override
+	public ItemStack getStateStack() {
+		return projectedBlockSlot.getStack();
+	}
+
+	@Override
+	public IBlockState getSavedState() {
+		return te.getProjectedState();
 	}
 }