+++ src/main/java/net/geforcemods/securitycraft/renderers/TileEntityProjectorRenderer.java
@@ -2,12 +2,14 @@ package net.geforcemods.securitycraft.renderers;
 
 import net.geforcemods.securitycraft.blocks.BlockProjector;
 import net.geforcemods.securitycraft.tileentity.TileEntityProjector;
+import net.geforcemods.securitycraft.util.TileEntityRenderDelegate;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.renderer.BlockRendererDispatcher;
 import net.minecraft.client.renderer.GlStateManager;
 import net.minecraft.client.renderer.texture.TextureMap;
 import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
+import net.minecraft.util.EnumBlockRenderType;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
 import net.minecraftforge.fml.relauncher.Side;
@@ -17,6 +19,8 @@ import net.minecraftforge.fml.relauncher.SideOnly;
 public class TileEntityProjectorRenderer extends TileEntitySpecialRenderer<TileEntityProjector> {
 	@Override
 	public void render(TileEntityProjector te, double x, double y, double z, float partialTicks, int destroyStage, float alpha) {
+		TileEntityRenderDelegate.DISGUISED_BLOCK.tryRenderDelegate(te, x, y, z, partialTicks, destroyStage, alpha);
+
 		if (te.isActive() && !te.isEmpty()) {
 			EnumFacing facing = te.getWorld().getBlockState(te.getPos()).getValue(BlockProjector.FACING);
 
@@ -37,11 +41,15 @@ public class TileEntityProjectorRenderer extends TileEntitySpecialRenderer<TileE
 
 					if (pos != null && te.getWorld().isAirBlock(pos)) {
 						BlockRendererDispatcher blockRendererDispatcher = Minecraft.getMinecraft().getBlockRendererDispatcher();
-						IBlockState state = te.getProjectedBlock().getStateFromMeta(te.getStackInSlot(36).getMetadata());
+						IBlockState state = te.getProjectedState();
 
 						GlStateManager.disableCull();
 						GlStateManager.scale(0.9999D, 0.9999D, 0.9999D); //counteract z-fighting between fake blocks
-						blockRendererDispatcher.renderBlockBrightness(state, te.getWorld().getLightBrightness(pos));
+
+						if (state.getRenderType() == EnumBlockRenderType.MODEL)
+							blockRendererDispatcher.renderBlockBrightness(state, te.getWorld().getLightBrightness(pos));
+
+						TileEntityRenderDelegate.PROJECTOR.tryRenderDelegate(te, 0, 0, -1, partialTicks, destroyStage, alpha);
 						GlStateManager.enableCull();
 					}